+++ src/main/java/basakan/fryday/controller/category/response/CategoryResponse.java
@@ -1,7 +1,7 @@
-package basakan.fryday.controller.dto;
+package basakan.fryday.controller.category.response;
 
-import basakan.fryday.domain.Category;
-import basakan.fryday.domain.CategoryColor;
+import basakan.fryday.domain.category.Category;
+import basakan.fryday.domain.todo.CategoryColor;
 import lombok.Getter;
 
 @Getter
