+++ src/main/java/basakan/fryday/controller/todo/TodoController.java
@@ -1,7 +1,14 @@
-package basakan.fryday.controller;
+package basakan.fryday.controller.todo;
 
-import basakan.fryday.controller.dto.*;
 import basakan.fryday.common.response.ApiResponse;
+import basakan.fryday.controller.todo.request.MemoRequest;
+import basakan.fryday.controller.dto.OrderUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoDateUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoSaveRequest;
+import basakan.fryday.controller.todo.response.CharacterStatusResponse;
+import basakan.fryday.controller.todo.response.MemoResponse;
+import basakan.fryday.controller.todo.response.TodoListResponse;
+import basakan.fryday.controller.todo.response.TodoResponse;
 import basakan.fryday.service.TodoService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
