+++ src/main/java/basakan/fryday/controller/todo/request/TodoDateUpdateRequest.java
@@ -1,6 +1,5 @@
-package basakan.fryday.controller.dto;
+package basakan.fryday.controller.todo.request;
 
-import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
