+++ src/main/java/basakan/fryday/domain/category/Category.java
@@ -1,5 +1,7 @@
-package basakan.fryday.domain;
+package basakan.fryday.domain.category;
 
+import basakan.fryday.domain.BaseEntity;
+import basakan.fryday.domain.todo.CategoryColor;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -11,7 +13,7 @@ import java.time.LocalDateTime;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Category extends BaseEntity{
+public class Category extends BaseEntity {
 
     @Column(nullable = false)
     private String name;
