+++ src/main/java/basakan/fryday/domain/todo/CharacterStatus.java
@@ -1,4 +1,4 @@
-package basakan.fryday.domain;
+package basakan.fryday.domain.todo;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
