+++ src/main/java/basakan/fryday/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package basakan.fryday.repository;
 
-import basakan.fryday.domain.Category;
+import basakan.fryday.domain.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
