+++ src/main/java/basakan/fryday/service/TodoService.java
@@ -2,11 +2,18 @@ package basakan.fryday.service;
 
 import basakan.fryday.common.ErrorCode;
 import basakan.fryday.common.exception.BusinessException;
-import basakan.fryday.controller.dto.*;
+import basakan.fryday.controller.todo.request.MemoRequest;
+import basakan.fryday.controller.dto.OrderUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoDateUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoSaveRequest;
+import basakan.fryday.controller.todo.response.CharacterStatusResponse;
+import basakan.fryday.controller.todo.response.MemoResponse;
+import basakan.fryday.controller.todo.response.TodoListResponse;
+import basakan.fryday.controller.todo.response.TodoResponse;
 import basakan.fryday.domain.BaseEntity;
-import basakan.fryday.domain.Category;
-import basakan.fryday.domain.CharacterStatus;
-import basakan.fryday.domain.Todo;
+import basakan.fryday.domain.category.Category;
+import basakan.fryday.domain.todo.CharacterStatus;
+import basakan.fryday.domain.todo.Todo;
 import basakan.fryday.repository.CategoryRepository;
 import basakan.fryday.repository.TodoRepository;
 import lombok.RequiredArgsConstructor;
