+++ src/test/java/basakan/fryday/controller/CategoryControllerTest.java
@@ -1,9 +1,14 @@
 package basakan.fryday.controller;
 
 import basakan.fryday.RestDocsSupport;
-import basakan.fryday.controller.dto.*;
-import basakan.fryday.domain.Category;
-import basakan.fryday.domain.CategoryColor;
+import basakan.fryday.controller.category.CategoryController;
+import basakan.fryday.controller.category.request.CategoryCreateRequest;
+import basakan.fryday.controller.category.request.CategoryUpdateRequest;
+import basakan.fryday.controller.dto.OrderUpdateRequest;
+import basakan.fryday.controller.category.response.CategoryReadResponse;
+import basakan.fryday.controller.category.response.CategoryResponse;
+import basakan.fryday.domain.category.Category;
+import basakan.fryday.domain.todo.CategoryColor;
 import basakan.fryday.service.CategoryService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
