+++ src/test/java/basakan/fryday/controller/TodoControllerTest.java
@@ -3,11 +3,19 @@ package basakan.fryday.controller;
 import basakan.fryday.RestDocsSupport;
 import basakan.fryday.common.ErrorCode;
 import basakan.fryday.common.exception.BusinessException;
-import basakan.fryday.controller.dto.*;
-import basakan.fryday.domain.Category;
-import basakan.fryday.domain.CategoryColor;
-import basakan.fryday.domain.CharacterStatus;
-import basakan.fryday.domain.Todo;
+import basakan.fryday.controller.todo.request.MemoRequest;
+import basakan.fryday.controller.dto.OrderUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoDateUpdateRequest;
+import basakan.fryday.controller.todo.request.TodoSaveRequest;
+import basakan.fryday.controller.todo.response.CharacterStatusResponse;
+import basakan.fryday.controller.todo.response.MemoResponse;
+import basakan.fryday.controller.todo.response.TodoListResponse;
+import basakan.fryday.controller.todo.response.TodoResponse;
+import basakan.fryday.controller.todo.TodoController;
+import basakan.fryday.domain.category.Category;
+import basakan.fryday.domain.todo.CategoryColor;
+import basakan.fryday.domain.todo.CharacterStatus;
+import basakan.fryday.domain.todo.Todo;
 import basakan.fryday.service.TodoService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
